Box fixes, javascript error, and pipeline cleanup
Box connection now uses threads to more quickly archive frames and json. Also goes to the correct sub-folders instead of dumping in root. Also adds a date-hour timestamp to the files to avoid duplicate 'frame_001' errors. Also updated the javascript of card creation to avoid a constant variable assignment error.

Still need to figure out the work order cards not displaying a URL to Salesforce and only showing object [object] at the bottom of the card.