Work Order Card updates
Hopefully fixed the cards so that they display the ai analysis text as strings and include a button to 'View in Salesforce'. It might make the cards a bit too tall. But I also messed with the css to maybe make the cards a little wider and less shy to be near other cards. So we'll see