Mode Update (timelapse, video)
Updated the system to take in a 'mode' argument at run_headless or main(__main__): 'timelapse' or 'video'. If timelapse, it extracts and analyzes every frame of footage. If video, it takes the provided framerate (default 1) into account and only parses about 1/30th of the frames (depending on video fps).